+++ src/overrides/etc.js
@@ -3,5 +3,4 @@ export const etcRules = {
   'etc/no-const-enum': 'error',
   'etc/no-enum': 'error',
   'etc/no-t': 'error',
-  'etc/prefer-interface': 'error',
 }