+++ changelog.md
@@ -6,9 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Disable `etc/prefer-interface`
 
 ## [0.3.1] - 2023-03-30
-### Changed
+### Fixed
 - Upgrade dependencies
 - Add `jsx-runtime` config to react config
 
@@ -25,7 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fix auto detection of libraries
 - Fix react settings
-- turn off `unicorn/switch-case-braces`
+- Turn off `unicorn/switch-case-braces`
 
 
 ## [0.1.0] - 2023-03-06